# generated-code/src/main/java/org/openapitools/api/TransactionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import java.math.BigDecimal;
import org.openapitools.model.TransactionPage;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-06-27T12:36:29.253976+02:00[Europe/Oslo]", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "transactions", description = "the transactions API")
public interface TransactionsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /transactions/statuses/{status} : Get transactions by status
     *
     * @param status  (required)
     * @param offset  (optional)
     * @param limit  (optional)
     * @param xRequestID  (optional)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "transactionsStatusesStatusGet",
        summary = "Get transactions by status",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TransactionPage.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/transactions/statuses/{status}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<TransactionPage> transactionsStatusesStatusGet(
        @Parameter(name = "status", description = "", required = true, in = ParameterIn.PATH) @PathVariable("status") BigDecimal status,
        @Parameter(name = "offset", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "offset", required = false) @Nullable Integer offset,
        @Parameter(name = "limit", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false) @Nullable Integer limit,
        @Parameter(name = "X-Request-ID", description = "", in = ParameterIn.HEADER) @RequestHeader(value = "X-Request-ID", required = false) @Nullable String xRequestID
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"total\" : 2, \"offset\" : 5, \"count\" : 1, \"limit\" : 5, \"items\" : [ { \"amount\" : 0.8008281904610115, \"id\" : \"id\", \"status\" : 6.027456183070403 }, { \"amount\" : 0.8008281904610115, \"id\" : \"id\", \"status\" : 6.027456183070403 } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}