# target/classes/openapi/transaction-api.yaml
openapi: 3.0.3
info:
  title: Transaction Service API
  version: 1.0.0
  description: API to handle transactions

servers:
  - url: http://localhost:8081

paths:
  /transactions/statuses/{status}:
    get:
      summary: Get transactions by status
      operationId: getTransactionsByStatus
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: number
          description: Status code to filter transactions
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: X-Request-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPage'
        '404':
          description: No transactions found for the given status

  /transactions:
    post:
      summary: Create a new transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDTO'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '400':
          description: Bad request
        '500':
          description: Internal server error

components:
  schemas:
    TransactionDTO:
      type: object
      required:
        - amount
        - status
      properties:
        id:
          type: string
          description: Unique identifier of the transaction
        amount:
          type: number
          format: double
          description: Amount involved in the transaction
        status:
          type: number
          description: Status of the transaction

    TransactionPage:
      type: object
      properties:
        total:
          type: integer
          example: 100
        offset:
          type: integer
          example: 0
        limit:
          type: integer
          example: 10
        count:
          type: integer
          example: 10
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDTO'